Change class Product

+++ Who-Owes-Money/Migrations/20210919162123_V1.cs
@@ -53,7 +53,7 @@ namespace Who_Owes_Money.Migrations
                     Id = table.Column<int>(nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     ProductName = table.Column<string>(maxLength: 20, nullable: false),
-                    Price = table.Column<int>(nullable: false),
+                    Price = table.Column<double>(nullable: false),
                     UserName = table.Column<string>(nullable: true)
                 },
                 constraints: table =>
